requirements.txt
fastapi
uvicorn
torch
torchvision
numpy
pillow
jinja2
python-multipart
requests